+++ game_patch/multi/server_internal.h
@@ -372,15 +372,11 @@ struct AlpineRestrictConfig
 {
     bool advertise_alpine = true;
     bool only_welcome_alpine = false;
-    bool reject_incompatible_clients = true;
     bool clients_require_alpine = false;
 
     // below options are only used if clients_require_alpine = true
     bool reject_non_alpine_clients = false;
-    bool alpine_server_version_enforce_min = false;
     bool alpine_require_release_build = false;
-    VoteConfig vote_match;
-    OvertimeConfig overtime;
 };
 
 struct SpawnLifeConfig
@@ -654,6 +650,8 @@ struct AlpineServerConfig
     InactivityConfig inactivity_config;
     DamageNotificationConfig damage_notification_config;
     ClickLimiterConfig click_limiter_config;
+    VoteConfig vote_match;
+    OvertimeConfig overtime;
     VoteConfig vote_kick;
     VoteConfig vote_level;
     VoteConfig vote_gametype;